Split up fingerprint info component
