+++ frontend/src/metabase/components/MetadataInfo/ColumnFingerprintInfo/CategoryFingerprint.styled.tsx
