+++ frontend/src/metabase/components/MetadataInfo/ColumnFingerprintInfo/CategoryFingerprint.unit.spec.js
