+++ frontend/src/metabase/components/MetadataInfo/ColumnFingerprintInfo/ColumnFingerprintInfo.unit.spec.js
@@ -7,7 +7,7 @@ import {
 } from "metabase-types/api/mocks/presets";
 import { createMockState } from "metabase-types/store/mocks";
 import Dimension from "metabase-lib/Dimension";
-import FieldFingerprintInfo from "./FieldFingerprintInfo";
+import { TableColumnFingerprintInfo } from "./ColumnFingerprintInfo";
 
 const state = createMockState({
   entities: createMockEntitiesState({
@@ -19,7 +19,7 @@ const metadata = getMetadata(state);
 function setup(field) {
   return renderWithProviders(
     <div data-testid="container">
-      <FieldFingerprintInfo
+      <TableColumnFingerprintInfo
         field={field}
         timezone={field.table?.database?.timezone}
       />
