+++ frontend/src/metabase/components/MetadataInfo/ColumnFingerprintInfo/NumberFingerprint.tsx
