+++ frontend/src/metabase/components/MetadataInfo/ColumnFingerprintInfo/index.ts
@@ -1,0 +1,1 @@
+export * from "./ColumnFingerprintInfo";
