--- frontend/src/metabase/components/MetadataInfo/FieldFingerprintInfo/index.ts
@@ -1,1 +1,0 @@
-export * from "./FieldFingerprintInfo";
